// src/app/authentication/interfaces/user-response.interface.ts
// Generated by https://quicktype.io

export interface UserResponse {
    message: string;
    user:    User;
}

export interface User {
    links:     Link[];
    _id:       string;
    fullname:  string;
    username:  string;
    password:  string;
    createdAt: string;
    updatedAt: string;
}

export interface Link {
    _id:         string;
    title:       string;
    url:         string;
    description: string;
    createdAt:   string;
    updatedAt:   string;
}